requirements.txt
torch
transformers
pillow
numpy
pandas
scikit-learn
tqdm
